requirements.txt
openai==1.68.2
python-dotenv
streamlit==1.44.0